requirements.txt
python-telegram-bot==20.0a2
gspread==5.4.0
matplotlib==3.5.2
